refactor: RTP GLM5 tests merge
